Runtime annotation grammar updated (N#k)
The runtime rule N#k was implemented as N%k. Now the grammar is updated and the use of this rule is according to GODA's article. The rule N#k is working properly and N%k no longer works.

RTParser and RTRegex now force antlr to throw a grammar error instead of handling it.